requirements.txt
openai>=1.51.0
pydantic>=2.5.0
python-dotenv>=1.0.0
streamlit>=1.28.0
pandas>=2.1.0
tokencost>=0.1.12
streamlit-nested-layout>=0.1.0
tenacity>=8.2.2
diff-match-patch>=20230430
google-generativeai>=0.8.3
anthropic>=0.36.1